VoicerNode:shouldSteal checks latency window only if synth exists
Complicated. You should steal if the node is needed and a synth
exists and a/ it's playing or b/ it isn't playing yet (because of
latency) but it *will* be playing when the current trigger request
processes in the server. Previously it was checking b/ even if
there's no synth (but if there's no synth, you never need to steal).