Voicer: Keep globalControls in new ValidatingIdentityDictionary class
It is too easy to do: aVoicer.globalControls[\name] = 3500
... then, kablooie anywhere that expects to be able to iterate
over actual global controls.

The ValidatingIdentityDictionary functions 1/ reject anything other
than a VoicerGlobalControl to be put into it and 2/ if you try to
put a number in, it will set the control to that value (because this
is probably what you wanted, and in a live-coding situation, just
do something nice rather than throw an ugly error up on screen).